backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pandas==2.2.2
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
redis==5.0.1
scikit-learn==1.4.2
statsmodels==0.14.2
prophet==1.1.5
pyproj==3.6.1
boto3==1.34.144
statsmodels>=0.14.0
prophet>=1.1.4
scikit-learn>=1.3.0
